[Privatization] Add ParentPropertyLookupGuard service
