+++ rules/Privatization/Guard/ParentPropertyLookupGuard.php
@@ -1,0 +1,123 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Rector\Privatization\Guard;
+
+use PhpParser\Node;
+use PhpParser\Node\Expr\PropertyFetch;
+use PhpParser\Node\Expr\StaticPropertyFetch;
+use PhpParser\Node\Expr\Variable;
+use PhpParser\Node\Stmt;
+use PhpParser\Node\Stmt\Class_;
+use PhpParser\Node\Stmt\Property;
+use PHPStan\Reflection\ClassReflection;
+use Rector\Core\Enum\ObjectReference;
+use Rector\Core\NodeAnalyzer\PropertyFetchAnalyzer;
+use Rector\Core\PhpParser\AstResolver;
+use Rector\Core\PhpParser\Node\BetterNodeFinder;
+use Rector\Core\Reflection\ReflectionResolver;
+use Rector\NodeNameResolver\NodeNameResolver;
+
+final class ParentPropertyLookupGuard
+{
+    public function __construct(
+        private readonly BetterNodeFinder $betterNodeFinder,
+        private readonly ReflectionResolver $reflectionResolver,
+        private readonly NodeNameResolver $nodeNameResolver,
+        private readonly PropertyFetchAnalyzer $propertyFetchAnalyzer,
+        private readonly AstResolver $astResolver
+    ) {
+    }
+
+    public function isLegal(Property $property): bool
+    {
+        $class = $this->betterNodeFinder->findParentType($property, Class_::class);
+
+        if (! $class instanceof Class_) {
+            return false;
+        }
+
+        if ($class->extends === null) {
+            return false;
+        }
+
+        $classReflection = $this->reflectionResolver->resolveClassReflection($property);
+        if (! $classReflection instanceof ClassReflection) {
+            return false;
+        }
+
+        $propertyName = $this->nodeNameResolver->getName($property);
+        $className = $classReflection->getName();
+
+        foreach ($classReflection->getParents() as $parentClassReflection) {
+            if ($parentClassReflection->hasProperty($propertyName)) {
+                return false;
+            }
+
+            if ($this->isFoundInParentClassMethods($parentClassReflection, $propertyName, $className)) {
+                return false;
+            }
+        }
+
+        return true;
+    }
+
+    private function isFoundInParentClassMethods(
+        ClassReflection $parentClassReflection,
+        string $propertyName,
+        string $className
+    ): bool {
+        $classLike = $this->astResolver->resolveClassFromName($parentClassReflection->getName());
+        if (! $classLike instanceof Class_) {
+            return false;
+        }
+
+        $methods = $classLike->getMethods();
+        foreach ($methods as $method) {
+            $isFound = $this->isFoundInMethodStmts((array) $method->stmts, $propertyName, $className);
+            if ($isFound) {
+                return true;
+            }
+        }
+
+        return false;
+    }
+
+    /**
+     * @param Stmt[] $stmts
+     */
+    private function isFoundInMethodStmts(array $stmts, string $propertyName, string $className): bool
+    {
+        return (bool) $this->betterNodeFinder->findFirst($stmts, function (Node $subNode) use (
+            $propertyName,
+            $className
+        ): bool {
+            if (! $this->propertyFetchAnalyzer->isPropertyFetch($subNode)) {
+                return false;
+            }
+
+            /** @var PropertyFetch|StaticPropertyFetch $subNode */
+            if ($subNode instanceof PropertyFetch) {
+                if (! $subNode->var instanceof Variable) {
+                    return false;
+                }
+
+                if (! $this->nodeNameResolver->isName($subNode->var, 'this')) {
+                    return false;
+                }
+
+                return $this->nodeNameResolver->isName($subNode, $propertyName);
+            }
+
+            if (! $this->nodeNameResolver->isNames(
+                $subNode->class,
+                [ObjectReference::SELF()->getValue(), ObjectReference::STATIC()->getValue(), $className]
+            )) {
+                return false;
+            }
+
+            return $this->nodeNameResolver->isName($subNode->name, $propertyName);
+        });
+    }
+}
